Refactored loading of  mesh/solution, separating stream and file loaders.
