+++ makefile
@@ -252,10 +252,11 @@ Ccc  = $(strip $(CC) $(CFLAGS) $(GL_OPTS))
 ALL_SOURCE_FILES = \
  lib/gl/renderer.cpp lib/gl/renderer_core.cpp lib/gl/renderer_ff.cpp \
  lib/gl/shader.cpp lib/gl/types.cpp lib/aux_js.cpp lib/aux_vis.cpp \
- lib/font.cpp lib/gl2ps.c lib/gltf.cpp lib/material.cpp lib/openglvis.cpp \
- lib/palettes.cpp lib/palettes_base.cpp lib/sdl.cpp lib/sdl_helper.cpp lib/sdl_main.cpp \
- lib/stream_reader.cpp lib/threads.cpp lib/vsdata.cpp lib/vssolution.cpp \
- lib/vssolution3d.cpp lib/vsvector.cpp lib/vsvector3d.cpp
+ lib/data_state.cpp lib/file_reader.cpp lib/font.cpp lib/gl2ps.c lib/gltf.cpp \
+ lib/material.cpp lib/openglvis.cpp lib/palettes.cpp lib/palettes_base.cpp \
+ lib/sdl.cpp lib/sdl_helper.cpp lib/sdl_main.cpp lib/stream_reader.cpp \
+ lib/threads.cpp lib/vsdata.cpp lib/vssolution.cpp lib/vssolution3d.cpp \
+ lib/vsvector.cpp lib/vsvector3d.cpp
 OBJC_SOURCE_FILES = $(if $(NOTMAC),,lib/sdl_mac.mm)
 DESKTOP_ONLY_SOURCE_FILES = \
  lib/gl/renderer_ff.cpp lib/threads.cpp lib/gl2ps.c lib/sdl_x11.cpp
@@ -269,9 +270,10 @@ COMMON_SOURCE_FILES = $(filter-out \
 HEADER_FILES = \
  lib/gl/attr_traits.hpp lib/gl/platform_gl.hpp lib/gl/renderer.hpp \
  lib/gl/shader.hpp lib/gl/renderer_core.hpp lib/gl/renderer_ff.hpp \
- lib/gl/types.hpp lib/aux_vis.hpp lib/font.hpp lib/geom_utils.hpp lib/gl2ps.h \
- lib/logo.hpp lib/material.hpp lib/openglvis.hpp lib/palettes.hpp lib/palettes_base.hpp \
- lib/sdl.hpp lib/sdl_helper.hpp lib/sdl_mac.hpp lib/sdl_main.hpp lib/sdl_x11.hpp \
+ lib/gl/types.hpp lib/aux_vis.hpp lib/data_state.hpp lib/file_reader.hpp \
+ lib/font.hpp lib/geom_utils.hpp lib/gl2ps.h lib/logo.hpp lib/material.hpp \
+ lib/openglvis.hpp lib/palettes.hpp lib/palettes_base.hpp lib/sdl.hpp \
+ lib/sdl_helper.hpp lib/sdl_mac.hpp lib/sdl_main.hpp lib/sdl_x11.hpp \
  lib/stream_reader.hpp lib/threads.hpp lib/visual.hpp lib/vsdata.hpp \
  lib/vssolution.hpp lib/vssolution3d.hpp lib/vsvector.hpp lib/vsvector3d.hpp
